The preview scene's worker count is unsigned, like the value it holds
sceneworkers::count() returns unsigned, and for Tier::MainThread the value it
returns is kSceneMainThreadOnly = ~0u — a MODE, not a small count: the backend
spawns no thread and runs every parallel pass inline. The thumbnail renderer is
on that tier, so ~0u is not a hypothetical here.

The member stored it as int and cast both ways, which makes the ONE value that
carries meaning the one value the round trip is implementation-defined for.
Store what count() returns.

+++ src/bridge/enginepreviewscene.cpp
@@ -24,7 +24,7 @@ EnginePreviewScene::EnginePreviewScene(const std::shared_ptr<Engine> &engine,
                                        const char *namePrefix, sceneworkers::Tier tier)
     : mEngine(engine)
     , mNamePrefix(namePrefix)
-    , mWorkerThreads(int(sceneworkers::count(tier)))
+    , mWorkerThreads(sceneworkers::count(tier))
 {
 }
 
@@ -53,7 +53,7 @@ bool EnginePreviewScene::attach(View *view)
         if (mView) mView->setScene(nullptr);
     } else if (!mScene) {
         // ORDER: the View above already exists, so the Scene can be created.
-        mScene = engine->createScene(uniqueName(mNamePrefix, this), unsigned(mWorkerThreads));
+        mScene = engine->createScene(uniqueName(mNamePrefix, this), mWorkerThreads);
         if (!mScene) return false;
         configureScene(mScene);
         mMirror.reset(new SceneMirror(mScene));
